// apps/mobile/src/api/generated/experienceState.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Hyperlocal Kitchen API
 * OpenAPI spec version: 0.1.0
 */
import type { ExperienceStateMode } from './experienceStateMode';
import type { KitchenOwnOut } from './kitchenOwnOut';

export interface ExperienceState {
  mode?: ExperienceStateMode;
  owned_kitchen?: KitchenOwnOut | null;
}